requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
azure-cognitiveservices-speech==1.41.1
    # via -r requirements/requirements.in
azure-core==1.32.0
    # via azure-storage-blob
azure-storage-blob==12.23.1
    # via -r requirements/requirements.in
build==1.2.2.post1
    # via pip-tools
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
cryptography==43.0.3
    # via azure-storage-blob
idna==3.10
    # via requests
isodate==0.7.2
    # via azure-storage-blob
packaging==24.2
    # via build
pip-tools==7.4.1
    # via -r requirements/requirements.in
pycparser==2.22
    # via cffi
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.0.1
    # via -r requirements/requirements.in
requests==2.32.3
    # via
    #   -r requirements/requirements.in
    #   azure-core
six==1.16.0
    # via azure-core
tomli==2.0.2
    # via
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-storage-blob
urllib3==2.2.3
    # via requests
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
